requirements.txt
Flask==1.0.2
pandas==0.23.4
mtgsdk==1.3.0
beautifulsoup4==4.6.3
requests2==2.16.0
mlxtend==0.14.0
